deep_excavation/services/geometry/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 几何处理
gmsh==4.11.1
numpy==1.24.3

# 微服务基础设施
python-consul==1.1.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
jaeger-client==4.8.0

# 消息队列
pika==1.3.2

# 缓存
redis==5.0.1

# 日志
structlog==23.2.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1 